moved components of  pareto and balance under charts
