+++ src/app/balance/page.tsx
@@ -1,7 +1,7 @@
 import { getBalance } from "@/lib/services/balance";
 
 import protectedPage from "@/lib/protectedPage";
-import CliBalancePage from "@/components/balance";
+import CliBalancePage from "@/components/charts/balance";
 
 export default protectedPage({
   getData: getBalance,
