+++ src/components/charts/pareto/chart.tsx
