+++ src/components/charts/pareto/tooltip.tsx
